webapp/requirements.txt
flask
flask-wtf
flask-bootstrap
folium
geopandas
psycopg2-binary
